Adapted to Student controller to the student routes. below is a list of modifications and fixes made to get needed features in place: - Rewrite of all Student routes to use int student ctrl. - Removal of roomId from student routes, as the new controller allows them to be more decoupled - addRequest route util updated to adapt to SubSchemaController objects - New setRoom.fromStudent middleware method created - Add new SubSchemaController.findOwner method to allow a subdocument to reverse lookup it's owner document - Fixed the SchemaController.deleteOne method to call model.findByIdAndDelete instead of model.deleteOne - StudentController.createOne updated to correctly use the createdBy option - Updated the createStudentValidation to expect a roomId option instead of belongsTo - Added new hookable functionality to ctrl option organization to allows readied options to be filtered before creating the final route

+++ app/controllers/definitions/StudentController.js
@@ -1,3 +1,4 @@
+const ObjectId = require("mongoose").Types.ObjectId;
 const SubSchemaController = require("../types/SubSchemaController");
 
 const roomCtrl = require("../room");
@@ -17,7 +18,7 @@ const feedCtrl = require("../feed");
  * @typedef {CreateSubDocOptions & CreateStudentOptionsData} CreateStudentOptions
  */
 
-const makeFeed = async data => {
+const makeFeed = async (data, createdBy) => {
 
     const feed = await feedCtrl.createOne({ data }, { save: false });
 
@@ -58,7 +59,7 @@ class StudentController extends SubSchemaController {
             room: belongsTo,
             for: newStudent._id,
             in: "students"
-        });
+        }, createdBy._id);
 
         return newStudent;
 
@@ -78,7 +79,7 @@ class StudentController extends SubSchemaController {
 
         const deletedStudent = await super.deleteOne({ docId });
 
-        await feedCtrl.deleteOne({ docId: deletedStudent.feed });
+        if( deletedStudent.feed ) await feedCtrl.deleteOne({ docId: deletedStudent.feed });
 
         return deletedStudent;
